Refactor SV1 and SV2 modules for improved structure and functionality
- Renamed and reorganized modules for clarity, including the introduction of `sv1` and `sv2` directories.
- Added new `Downstream` and `ChannelManager` implementations to enhance message handling and processing.
- Introduced new message handler logic for upstream and downstream communication.
- Implemented share validation and extended mining channel management.
- Improved logging throughout the modules for better traceability and debugging.
