+++ roles/new-tproxy/src/lib/sv2/channel_manager/mod.rs
@@ -1,4 +1,3 @@
 pub mod channel_manager;
 pub mod message_handler;
-pub mod sv1_server;
 pub use channel_manager::ChannelManager;
