+++ roles/new-tproxy/src/lib/sv2/mod.rs
@@ -1,0 +1,6 @@
+pub mod channel_manager;
+pub mod upstream;
+
+pub use channel_manager::channel_manager::ChannelManager;
+pub use channel_manager::channel_manager::ChannelMappingMode;
+pub use upstream::upstream::Upstream;
